Update promise chaining syntax
The previous syntax was causing issues when the thePromise.finally() block was reached. It was throwing a "Query has already been executed" error. This is because thePromise.then() actually executes the promise, so by the time thePromise.finally() was reached, the query had already been executed. By chaining the .catch and .finally to the .then, we can avoid this issue while accomplishing the same goal.